pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "email-parser"
version = "1.0.0"
description = "AI-powered email parser with automated account setup"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Email Parser Team", email = "support@emailparser.ai"},
]
keywords = ["email", "parser", "ai", "gmail", "outlook", "llm", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Communications :: Email",
    "Topic :: Office/Business",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "msal>=1.20.0",
    "requests>=2.25.0",
    "pandas>=1.3.0",
    "tqdm>=4.60.0",
    "pydantic>=2.0.0",
    "python-dotenv>=0.19.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
    "beautifulsoup4>=4.10.0",
    "instructor>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "openpyxl>=3.0.0",
    "langgraph>=0.2.0",
    "langmem>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
email-parser = "email_parser.main:main"
email-parser-setup = "email_parser.setup.account_setup:main_setup_cli"

[project.urls]
Homepage = "https://github.com/your-org/email-parser"
Repository = "https://github.com/your-org/email-parser"
Documentation = "https://docs.emailparser.ai"
"Bug Reports" = "https://github.com/your-org/email-parser/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true